Don't reset notification time stamp for notifications with same ID
When a notification is sent with the same ID as an earlier one, the idea
is that this new notification more or less transparently replaces the
old one. However, when dunst has already started showing the "age" of
the old notification, the notification's timestamp will be reset, which
causes the age to be reset as well.
With this change we adjust this behavior, making sure that we keep the
original time stamp for what effectively is meant to be the same
notification, and, thus, stop resetting the age display.